add state clarification to this project
Thanks for creating this project. It looks awesome And I hope it may inspire us well. I also wish you the best. However, I think it can help many to ensure they know that for now this project is only pretending. It would be a shame that your work would be labeled as hype while all it might be is that you’re still in early development.

Best of luck and keep it up!