Update README.md
**Issue: Difficulty with Initial Setup and Clarity**

Users are finding it challenging to get started due to unclear instructions. Some common hurdles include not knowing where the terminal is, how to access it, and what needs to be downloaded beforehand. Along with setup.bat changed to ./setup.bat

**Proposed Solution**

To improve the clarity and ease of the initial setup process, we could provide more detailed, step-by-step instructions with prerequisite downloads and how to access and use the terminal.